pyproject.toml
[project]
name = "maris"
version = "0.1.0"
description = "Marine Asset Risk Intelligence System - Semantica POC"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core Framework (provides most functionality)
    "semantica==0.2.5",
    
    # LLM Provider for extraction
    "groq>=0.4.0",
    
    # Graph Database
    "falkordb>=1.0.0",
    
    # Essential utilities
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
maris = "maris.cli:main"
fetch-documents = "scripts.fetch_documents:fetch_documents"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools]
packages = ["maris", "scripts"]
include-package-data = true

[tool.hatchling.build.targets.wheel]
packages = ["maris"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
